refactor: move BackupCommand out of main
